+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- authorize operation-scoped diagnostics captures on Linux hosts using Yama restricted ptrace
+
 ## [0.4.1](https://github.com/mag1cfrog/delta-funnel/compare/delta-funnel-v0.4.0...delta-funnel-v0.4.1) - 2026-07-28
 
 ### Added